added search for museums

+++ src/main/java/com/museumsgide/demo/dto/MuseumDTO.java
@@ -11,13 +11,13 @@ public class MuseumDTO {
     private Long cityId;
     private Long catMuseumsId;
     private String description;
-    private Integer entrancePrice;
+    private Long entrancePrice;
 
-    public Integer getEntrancePrice() {
+    public Long getEntrancePrice() {
         return entrancePrice;
     }
 
-    public void setEntrancePrice(Integer entrancePrice) {
+    public void setEntrancePrice(Long entrancePrice) {
         this.entrancePrice = entrancePrice;
     }
